PAE-1652: Sonar fixes
Splits report-service.js and mongodb.js into smaller, cohesively
scoped modules, and shrinks the get-detail handler by extracting its
waste-record-exclusion warning and resubmission-eligibility checks
into helper functions.
