+++ src/reports/application/report-service.js
@@ -1,21 +1,16 @@
-import Boom from '@hapi/boom'
-import { badRequest, conflict } from '#common/helpers/logging/cdp-boom.js'
 import { resolveDetailedMaterial } from '#domain/organisations/registration-utils.js'
 import { getOrsDetailsMap } from '#overseas-sites/application/get-ors-details-map.js'
 import { getIssuedTonnage } from '#packaging-recycling-notes/application/get-issued-tonnage.js'
 import { latestSubmittedSummaryLog } from '#waste-balances/application/latest-submitted-summary-log.js'
 import { wasteRecordStatesForHead } from '#waste-records/application/read-summary-log-row-states.js'
 import { aggregateReportDetail } from '#reports/domain/aggregation/aggregate-report-detail.js'
-import { generateAllPeriodsForYear } from '#reports/domain/generate-reporting-periods.js'
 import { getOperatorCategory } from '#reports/domain/operator-category.js'
-import { REPORT_STATUS } from '#reports/domain/report-status.js'
 import {
-  RESUBMISSION_INELIGIBLE_REASON,
-  isResubmissionRequired,
-  resubmissionIneligibleReasonToErrorCode
-} from '#reports/domain/resubmission.js'
-import { errorCodes } from '#reports/enums/error-codes.js'
-import { isClosedPeriodAdjustmentsEnabled } from '#root/config.js'
+  assertNoExistingReport,
+  assertResubmissionAllowed,
+  getValidatedPeriodInfo
+} from './create-report-validation.js'
+import { findReportIdBySubmissionNumber } from './submission-lookup.js'
 
 /**
  * @import { Registration, RegistrationAddress } from '#domain/organisations/registration.js'
@@ -49,67 +44,6 @@ export const createReportsService = (reportsRepository) => ({
 
   findPeriodicReports: (params) => reportsRepository.findPeriodicReports(params)
 })
-
-/**
- * Finds the submission summary for a specific submission number within periodic
- * reports, checking both the current slot and previous submissions.
- * @param {PeriodicReport[]} periodicReports
- * @param {number} year
- * @param {Cadence} cadence
- * @param {number} period
- * @param {number} submissionNumber
- * @returns {import('#reports/repository/port.js').ReportSummary | null}
- */
-function findSubmissionByNumber(
-  periodicReports,
-  year,
-  cadence,
-  period,
-  submissionNumber
-) {
-  const slot = periodicReports.find((pr) => pr.year === year)?.reports?.[
-    cadence
-  ]?.[period]
-  if (!slot) {
-    return null
-  }
-  if (slot.current?.submissionNumber === submissionNumber) {
-    return slot.current
-  }
-  return (
-    slot.previousSubmissions?.find(
-      (s) => s.submissionNumber === submissionNumber
-    ) ?? null
-  )
-}
-
-/**
- * Finds the report ID for a specific submission number within periodic reports,
- * checking both the current slot and previous submissions.
- * @param {PeriodicReport[]} periodicReports
- * @param {number} year
- * @param {Cadence} cadence
- * @param {number} period
- * @param {number} submissionNumber
- * @returns {string|null}
- */
-function findReportIdBySubmissionNumber(
-  periodicReports,
-  year,
-  cadence,
-  period,
-  submissionNumber
-) {
-  return (
-    findSubmissionByNumber(
-      periodicReports,
-      year,
-      cadence,
-      period,
-      submissionNumber
-    )?.id ?? null
-  )
-}
 
 /**
  * Looks up a stored report for a specific period and submission number.
@@ -152,225 +86,6 @@ export async function fetchReportBySubmissionNumber(
 }
 
 /**
- * Determines whether a submission is the latest submission for its period, by
- * submission number, regardless of status. Only the latest submission may be
- * unsubmitted: unsubmitting an earlier one would silently drop it from the admin
- * submission history (PAE-1657). This intentionally guards the absolute latest
- * submission, not merely the latest *submitted* one, so an earlier submitted
- * report cannot be unsubmitted while a later resubmission draft is in progress.
- * The caller separately requires the target to be submitted.
- * @param {import('#reports/repository/port.js').ReportsRepository} reportsRepository
- * @param {string} organisationId
- * @param {string} registrationId
- * @param {number} year
- * @param {Cadence} cadence
- * @param {number} period
- * @param {number} submissionNumber
- * @returns {Promise<boolean>}
- */
-export async function isLatestSubmission(
-  reportsRepository,
-  organisationId,
-  registrationId,
-  year,
-  cadence,
-  period,
-  submissionNumber
-) {
-  const periodicReports = await reportsRepository.findPeriodicReports({
-    organisationId,
-    registrationId
-  })
-  return isLatestSubmissionOf(
-    periodicReports,
-    year,
-    cadence,
-    period,
-    submissionNumber
-  )
-}
-
-/**
- * Whether `submissionNumber` is the highest submission number recorded for
- * its period, regardless of status — the same "absolute latest, not merely
- * latest submitted" rule {@link isLatestSubmission} enforces, extracted here
- * so it can run against a `periodicReports` list already fetched once by the
- * caller instead of each check re-fetching it.
- * @param {PeriodicReport[]} periodicReports
- * @param {number} year
- * @param {Cadence} cadence
- * @param {number} period
- * @param {number} submissionNumber
- * @returns {boolean}
- */
-function isLatestSubmissionOf(
-  periodicReports,
-  year,
-  cadence,
-  period,
-  submissionNumber
-) {
-  const slot = periodicReports.find((pr) => pr.year === year)?.reports?.[
-    cadence
-  ]?.[period]
-  const submissions = [
-    slot?.current,
-    ...(slot?.previousSubmissions ?? [])
-  ].flatMap((s) => (s ? [s.submissionNumber] : []))
-  if (submissions.length === 0) {
-    return false
-  }
-  return Math.max(...submissions) === submissionNumber
-}
-
-/**
- * Whether an operator may request resubmission on a submitted report
- * @param {PeriodicReport[]} periodicReports
- * @param {object} target
- * @param {import('#reports/domain/report-status.js').ReportStatus} target.status
- * @param {import('#reports/repository/port.js').ReportResubmissionRequired|null} [target.resubmissionRequired]
- * @param {number} target.year
- * @param {Cadence} target.cadence
- * @param {number} target.period
- * @param {number} target.submissionNumber
- * @returns {import('#reports/repository/port.js').ResubmissionIneligibleReason}
- */
-export function resubmissionEligibility(periodicReports, target) {
-  if (!isClosedPeriodAdjustmentsEnabled()) {
-    return RESUBMISSION_INELIGIBLE_REASON.FEATURE_DISABLED
-  }
-  if (target.status !== REPORT_STATUS.SUBMITTED) {
-    return RESUBMISSION_INELIGIBLE_REASON.NOT_SUBMITTED
-  }
-  if (isResubmissionRequired(target.resubmissionRequired)) {
-    return RESUBMISSION_INELIGIBLE_REASON.ALREADY_REQUESTED
-  }
-  const { year, cadence, period, submissionNumber } = target
-  if (
-    !isLatestSubmissionOf(
-      periodicReports,
-      year,
-      cadence,
-      period,
-      submissionNumber
-    )
-  ) {
-    return RESUBMISSION_INELIGIBLE_REASON.NOT_LATEST_SUBMISSION
-  }
-  return RESUBMISSION_INELIGIBLE_REASON.ELIGIBLE
-}
-
-/**
- * Whether an operator may request resubmission on the given submitted
- * report right now. Thin boolean wrapper over {@link resubmissionEligibility}.
- * @param {PeriodicReport[]} periodicReports
- * @param {object} target
- * @param {import('#reports/domain/report-status.js').ReportStatus} target.status
- * @param {import('#reports/repository/port.js').ReportResubmissionRequired|null} [target.resubmissionRequired]
- * @param {number} target.year
- * @param {Cadence} target.cadence
- * @param {number} target.period
- * @param {number} target.submissionNumber
- * @returns {boolean}
- */
-export function canRequestResubmission(periodicReports, target) {
-  return (
-    resubmissionEligibility(periodicReports, target) ===
-    RESUBMISSION_INELIGIBLE_REASON.ELIGIBLE
-  )
-}
-
-/**
- * Requests resubmission on an operator's own submitted report. Rejects
- * using {@link resubmissionEligibility}, then writes — the write can still
- * lose a race after the check passed, in which case it's rejected too.
- *
- * @param {object} params
- * @param {import('#reports/repository/port.js').ReportsRepository} params.reportsRepository
- * @param {string} params.organisationId
- * @param {string} params.registrationId
- * @param {number} params.year
- * @param {Cadence} params.cadence
- * @param {number} params.period
- * @param {number} params.submissionNumber
- * @param {import('#reports/repository/port.js').UserSummary} params.requestedBy
- * @returns {Promise<import('#reports/repository/port.js').MarkSubmittedReportRequiringResubmissionByOperatorFlaggedResult>}
- */
-export async function requestOperatorResubmission({
-  reportsRepository,
-  organisationId,
-  registrationId,
-  year,
-  cadence,
-  period,
-  submissionNumber,
-  requestedBy
-}) {
-  const reject = (reason) =>
-    conflict(
-      `Resubmission request not permitted for ${cadence} period ${period} of ${year} submission ${submissionNumber}`,
-      reason,
-      {
-        event: {
-          action: 'request_resubmission',
-          reason: `cadence=${cadence} period=${period} year=${year} submissionNumber=${submissionNumber} rejected=${reason}`
-        },
-        payload: { reason }
-      }
-    )
-
-  const periodicReports = await reportsRepository.findPeriodicReports({
-    organisationId,
-    registrationId
-  })
-
-  const target = findSubmissionByNumber(
-    periodicReports,
-    year,
-    cadence,
-    period,
-    submissionNumber
-  )
-
-  if (!target) {
-    throw Boom.notFound(
-      `No report found for ${cadence} period ${period} of ${year} submission ${submissionNumber}`
-    )
-  }
-
-  const eligibility = resubmissionEligibility(periodicReports, {
-    status: target.status,
-    resubmissionRequired: target.resubmissionRequired,
-    year,
-    cadence,
-    period,
-    submissionNumber
-  })
-
-  if (eligibility !== RESUBMISSION_INELIGIBLE_REASON.ELIGIBLE) {
-    throw reject(resubmissionIneligibleReasonToErrorCode(eligibility))
-  }
-
-  const result =
-    await reportsRepository.markSubmittedReportRequiringResubmissionByOperator({
-      organisationId,
-      registrationId,
-      year,
-      cadence,
-      period,
-      submissionNumber,
-      requestedBy,
-      requestedAt: new Date().toISOString()
-    })
-
-  if (!result) {
-    throw reject(errorCodes.resubmissionRequestPreconditionFailed)
-  }
-
-  return result
-}
-
-/**
  * Formats a site address object into a single-line string.
  * @param {RegistrationAddress|undefined} address
  * @returns {string|undefined}
@@ -382,185 +97,6 @@ function formatSiteAddress(address) {
   return [address.line1, address.line2, address.town, address.postcode]
     .filter(Boolean)
     .join(', ')
-}
-
-/**
- * @typedef {{ period: number, startDate: string, endDate: string, dueDate: string }} PeriodInfo
- */
-
-/**
- * Throws a 400 Boom with `output.payload.invalidPeriod` if the period isn't
- * in the cadence's valid range for the year.
- *
- * @param {number} period
- * @param {Cadence} cadence
- * @param {PeriodInfo[]} allPeriods
- * @returns {PeriodInfo}
- */
-const assertValidPeriod = (period, cadence, allPeriods) => {
-  const periodInfo = allPeriods.find((p) => p.period === period)
-  if (!periodInfo) {
-    const validPeriods = allPeriods.map((p) => p.period)
-    throw badRequest(
-      `Invalid period ${period} for cadence ${cadence}`,
-      errorCodes.invalidPeriod,
-      {
-        event: {
-          action: 'create_report',
-          reason: `actual=${period} cadence=${cadence} validPeriods=[${validPeriods.join(',')}]`
-        },
-        payload: { invalidPeriod: { actual: period, cadence, validPeriods } }
-      }
-    )
-  }
-  return periodInfo
-}
-
-/**
- * Throws a 400 Boom with `output.payload.periodNotEnded` if the period's
- * end date has not yet passed.
- *
- * @param {PeriodInfo} periodInfo
- * @param {number} period
- * @param {Cadence} cadence
- * @returns {void}
- */
-const assertPeriodEnded = (periodInfo, period, cadence) => {
-  const dayAfterEnd = new Date(periodInfo.endDate)
-  dayAfterEnd.setUTCDate(dayAfterEnd.getUTCDate() + 1)
-  if (dayAfterEnd > new Date()) {
-    const earliestSubmissionDate = dayAfterEnd.toISOString()
-    throw badRequest(
-      `Cannot create report for period ${period} — period has not yet ended`,
-      errorCodes.periodNotEnded,
-      {
-        event: {
-          action: 'create_report',
-          reason: `period=${period} cadence=${cadence} endDate=${periodInfo.endDate} earliestSubmissionDate=${earliestSubmissionDate}`
-        },
-        payload: {
-          periodNotEnded: {
-            period,
-            cadence,
-            endDate: periodInfo.endDate,
-            earliestSubmissionDate
-          }
-        }
-      }
-    )
-  }
-}
-
-/**
- * Throws a 409 Boom with `output.payload.existingReport` if a report for the
- * same period already exists.
- *
- * @param {PeriodicReport[]} periodicReports
- * @param {number} year
- * @param {Cadence} cadence
- * @param {number} period
- * @returns {void}
- */
-const assertNoExistingReport = (
-  periodicReports,
-  year,
-  cadence,
-  period,
-  submissionNumber
-) => {
-  const id = findReportIdBySubmissionNumber(
-    periodicReports,
-    year,
-    cadence,
-    period,
-    submissionNumber
-  )
-  if (id) {
-    throw conflict(
-      `Report already exists for ${cadence} period ${period} of ${year}`,
-      errorCodes.reportAlreadyExists,
-      {
-        event: {
-          action: 'create_report',
-          reason: `cadence=${cadence} period=${period} year=${year}`,
-          reference: id
-        },
-        payload: { existingReport: { id, cadence, period, year } }
-      }
-    )
-  }
-}
-
-/**
- * Throws a 409 Boom when creating a resubmission (submissionNumber > 1) is not
- * allowed. A resubmission requires the closed-period-adjustments feature flag
- * to be on and the previous submission to be a submitted report flagged as
- * requiring resubmission. The flag is checked first, then the block rule; each
- * failure carries a distinct `reason` so the frontend can tell them apart from
- * the duplicate-period conflict.
- *
- * @param {PeriodicReport[]} periodicReports
- * @param {number} year
- * @param {Cadence} cadence
- * @param {number} period
- * @param {number} submissionNumber
- * @returns {void}
- */
-const assertResubmissionAllowed = (
-  periodicReports,
-  year,
-  cadence,
-  period,
-  submissionNumber
-) => {
-  if (submissionNumber <= 1) {
-    return
-  }
-
-  const reject = (reason) =>
-    conflict(
-      `Resubmission ${submissionNumber} not permitted for ${cadence} period ${period} of ${year}`,
-      reason,
-      {
-        event: {
-          action: 'create_report',
-          reason: `cadence=${cadence} period=${period} year=${year} submissionNumber=${submissionNumber} rejected=${reason}`
-        },
-        payload: { reason }
-      }
-    )
-
-  if (!isClosedPeriodAdjustmentsEnabled()) {
-    throw reject(errorCodes.resubmissionFeatureDisabled)
-  }
-
-  const previous = findSubmissionByNumber(
-    periodicReports,
-    year,
-    cadence,
-    period,
-    submissionNumber - 1
-  )
-  const permitted =
-    previous?.status === REPORT_STATUS.SUBMITTED &&
-    isResubmissionRequired(previous?.resubmissionRequired)
-  if (!permitted) {
-    throw reject(errorCodes.resubmissionNotPermitted)
-  }
-}
-
-/**
- * Validates that a period exists for the given cadence and has ended.
- * @param {Cadence} cadence
- * @param {number} year
- * @param {number} period
- * @returns {{ startDate: string, endDate: string, dueDate: string }}
- */
-function getValidatedPeriodInfo(cadence, year, period) {
-  const allPeriods = generateAllPeriodsForYear(cadence, year)
-  const periodInfo = assertValidPeriod(period, cadence, allPeriods)
-  assertPeriodEnded(periodInfo, period, cadence)
-  return periodInfo
 }
 
 /**
