+++ src/reports/routes/get-detail.js
@@ -1,9 +1,7 @@
 import { StatusCodes } from 'http-status-codes'
 
-import {
-  fetchOrGenerateReportForPeriod,
-  canRequestResubmission
-} from '#reports/application/report-service.js'
+import { fetchOrGenerateReportForPeriod } from '#reports/application/report-service.js'
+import { canRequestResubmission } from '#reports/application/resubmission-service.js'
 import { periodParamsSchema, withRegistrationDetails } from './shared.js'
 import { reportDetailResponseSchema } from './response.schema.js'
 import { reportResponseFailAction } from './response-fail-action.js'
@@ -23,6 +21,72 @@ import { SCOPES } from '#common/helpers/auth/constants.js'
 
 export const reportsGetDetailPath =
   '/v1/organisations/{organisationId}/registrations/{registrationId}/reports/{year}/{cadence}/{period}/submissions/{submissionNumber}'
+
+/**
+ * @param {HapiRequest} request
+ * @param {import('#reports/domain/aggregation/aggregate-report-detail.js').AggregatedReportDetail | import('#reports/repository/port.js').Report} report
+ * @param {string} organisationId
+ * @param {string} registrationId
+ */
+function warnIfWasteRecordsExcluded(
+  request,
+  report,
+  organisationId,
+  registrationId
+) {
+  // The 'diagnostics' in report check acts as a type discriminator:
+  // fetchOrGenerateReportForPeriod returns Report | AggregatedReportDetail,
+  // and only AggregatedReportDetail carries diagnostics (and operatorCategory).
+  if (
+    !('diagnostics' in report) ||
+    report.diagnostics.wasteReceivedRecordsExcluded === 0
+  ) {
+    return
+  }
+
+  const { wasteReceivedRecordsExcluded } = report.diagnostics
+  request.logger.warn({
+    message:
+      'Waste records excluded from report due to mismatched date field — possible registered-only to accredited transition (ADR 0030)',
+    event: {
+      action: 'fetch_or_generate_report',
+      reason: `organisationId=${organisationId} registrationId=${registrationId} operatorCategory=${report.operatorCategory} wasteReceivedRecordsExcluded=${wasteReceivedRecordsExcluded}`
+    }
+  })
+}
+
+/**
+ * @param {ReportsRepository} reportsRepository
+ * @param {import('#reports/domain/aggregation/aggregate-report-detail.js').AggregatedReportDetail | import('#reports/repository/port.js').Report} report
+ * @param {string} organisationId
+ * @param {string} registrationId
+ */
+async function resolveCanRequestResubmission(
+  reportsRepository,
+  report,
+  organisationId,
+  registrationId
+) {
+  if ('diagnostics' in report) {
+    return false
+  }
+
+  const periodicReports = await reportsRepository.findPeriodicReports({
+    organisationId,
+    registrationId
+  })
+
+  return canRequestResubmission(periodicReports, {
+    status: report.status.currentStatus,
+    resubmissionRequired: report.resubmissionRequired,
+    year: report.year,
+    cadence: /** @type {import('#reports/domain/cadence.js').Cadence} */ (
+      report.cadence
+    ),
+    period: report.period,
+    submissionNumber: report.submissionNumber
+  })
+}
 
 export const reportsGetDetail = {
   method: 'GET',
@@ -89,43 +153,14 @@ export const reportsGetDetail = {
       submissionNumber
     })
 
-    // The 'diagnostics' in report check acts as a type discriminator:
-    // fetchOrGenerateReportForPeriod returns Report | AggregatedReportDetail,
-    // and only AggregatedReportDetail carries diagnostics (and operatorCategory).
-    if (
-      'diagnostics' in report &&
-      report.diagnostics.wasteReceivedRecordsExcluded > 0
-    ) {
-      const { wasteReceivedRecordsExcluded } = report.diagnostics
-      request.logger.warn({
-        message:
-          'Waste records excluded from report due to mismatched date field — possible registered-only to accredited transition (ADR 0030)',
-        event: {
-          action: 'fetch_or_generate_report',
-          reason: `organisationId=${organisationId} registrationId=${registrationId} operatorCategory=${report.operatorCategory} wasteReceivedRecordsExcluded=${wasteReceivedRecordsExcluded}`
-        }
-      })
-    }
+    warnIfWasteRecordsExcluded(request, report, organisationId, registrationId)
 
-    const canRequestResubmissionFlag = !('diagnostics' in report)
-      ? canRequestResubmission(
-          await reportsRepository.findPeriodicReports({
-            organisationId,
-            registrationId
-          }),
-          {
-            status: report.status.currentStatus,
-            resubmissionRequired: report.resubmissionRequired,
-            year: report.year,
-            cadence:
-              /** @type {import('#reports/domain/cadence.js').Cadence} */ (
-                report.cadence
-              ),
-            period: report.period,
-            submissionNumber: report.submissionNumber
-          }
-        )
-      : false
+    const canRequestResubmissionFlag = await resolveCanRequestResubmission(
+      reportsRepository,
+      report,
+      organisationId,
+      registrationId
+    )
 
     return h
       .response({
